Add files via upload
Cybele lab - det var noe som var i konflikt med hverandre
