+++ utils.py
@@ -14,6 +14,7 @@ torch.backends.cudnn.benchmark = True
 torch.backends.cudnn.deterministic = True
 
 
+
 def set_seed(seed: int):
     np.random.seed(seed)
     torch.manual_seed(seed)
@@ -104,3 +105,5 @@ def plot_loss(loss_dict: dict, label: str = None, npoints_to_average=1, plot_var
         steps, np.array(mean_loss) -
         np.array(loss_std), np.array(mean_loss) + loss_std,
         alpha=.2, label=f"{label} variance over {npoints_to_average} steps")
+    
+
